hospitalProject/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
torch
numpy
langchain
langchain-google-genai
langgraph
pytest
pytest-django
factory_boy
Faker
psycopg2-binary
requests
axios
gunicorn
whitenoise
dj-database-url
